% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_context.R
\name{h2o_context}
\alias{h2o_context}
\title{Get or create the H2OContext for a SparkContext}
\usage{
h2o_context(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}}
}
\description{
Get or create the H2OContext for a SparkContext
}

