\docType{package}
\name{sparklingwater}
\alias{sparklingwater}
\alias{sparklingwater-package}
\title{sparklingwater}
\format{A data frame with 21 rows and 13 variables}
\description{
sparklingwater

The Sparkling Water dataset from Léon Bottou's COS 405
lecture on PCA
}
\details{
\itemize{ \item Mineral.water Mineral water brand \item
Calcium \item Magnésium ... }
}
\keyword{sparklingwater}

